+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `OVERALL_PARAMS` dictionary
 - `__imbalancement_handler__` function
 - `vector_serializer` function
+- NPV micro/macro
 - 23 new distance/similarity
 	1. Dennis 
 	2. Digby